Udate Topological Sort
I had a version which I think is better then the one which was here. It also handles the case in which there are cycles in the graph, giving as output an empty list (= impossible scheduling). It also has a graph class which I think it's nice. I think it can be improved, I was actually curious if somebody had a better version of it.